update: fix return value from failed mkostemp
We'll probably never run into this on Linux, and it's even less likely
that mkostemp would fail, but we might as well account for it if we're
going to offer the fallback.

+++ src/update.c
@@ -379,7 +379,7 @@ static size_t write_handler(void *ptr, size_t size, size_t nmemb, void *data) {
 
 static int open_tmpfile(int flags) {
   const char *tmpdir;
-  char *p;
+  _cleanup_free_ char *p = NULL;
   int fd;
 
   tmpdir = getenv("TMPDIR");
@@ -399,8 +399,12 @@ static int open_tmpfile(int flags) {
   }
 
   fd = mkostemp(p, flags);
+  if (fd < 0) {
+    return -errno;
+  }
+
+  /* ignore any (unlikely) error */
   unlink(p);
-  free(p);
 
   return fd;
 }
